+++ R/cov_trait.R
@@ -3,11 +3,9 @@
 #' This function returns the expected covariance matrix of a trait vector simulated via `sim_trait`.
 #' Below there are `n` individuals.
 #'
+#' @inheritParams sim_trait
 #' @param kinship The `n`-by-`n` kinship matrix of the individuals.
 #' These values should be scaled such that an outbred individual has 1/2 self-kinship, the parent-child relationship is 1/4, etc (which is half the values sometimes defined for kinship).
-#' @param herit The heritability (proportion of trait variance due to genetics).
-#' @param sigma_sq Overall variance multiplicative factor (default 1).
-#' This factor corresponds to the variance of an outbred individual.
 #'
 #' @return The `n`-by-`n` trait covariance matrix equal to
 #' `sigma_sq * ( herit * 2 * kinship + ( 1 - herit ) * I )`,
@@ -27,8 +25,17 @@
 #' # covariance of simulated traits
 #' V <- cov_trait(kinship = kinship, herit = 0.8)
 #'
+#' @seealso
+#' [sim_trait()], [sim_trait_mvn()]
+#' 
 #' @export
-cov_trait <- function(kinship, herit, sigma_sq = 1) {
+cov_trait <- function(
+                      kinship,
+                      herit,
+                      sigma_sq = 1,
+                      labs = NULL,
+                      labs_sigma_sq = NULL
+                      ) {
     # construct V from kinship and herit
 
     # check for missing parameters
@@ -51,12 +58,40 @@ cov_trait <- function(kinship, herit, sigma_sq = 1) {
     if ( !isSymmetric( kinship ) )
         stop( '`kinship` must be a square, symmetric matrix!' )
     
+    # if there are group effects, check those
+    n_ind <- nrow( kinship )
+    labs <- check_labs( labs, labs_sigma_sq, n_ind, herit )
+    
     # identity matrix of same dimension as kinship
     I <- diag( nrow( kinship ) )
     # desired covariance matrix (except for overall scale)
     V <- 2 * herit * kinship + (1-herit) * I
-    # multiply by scale
+    
+    # if there are group effects, add them now
+    if ( !is.null( labs ) ) {
+        n_labs <- ncol( labs )
+        for ( i in 1L : n_labs ) {
+            # process level i
+            labs_i <- labs[ , i ]
+            labs_i_sigma_sq <- labs_sigma_sq[ i ]
+            # unique groups, implicitly numbered by order of appearance
+            groups_i <- unique( labs_i )
+            # map individuals to their groups by index
+            group_indexes_i <- match( labs_i, groups_i )
+            # in this level, we have a simple block matrix reflecting shared group membership
+            n_groups_i <- length( groups_i )
+            Ci <- matrix( 0, n_ind, n_ind )
+            for ( j in 1L : n_groups_i ) {
+                Ci <- Ci + tcrossprod( group_indexes_i == j )
+            }
+            # multiply by variance factor, and add to running covariance sum
+            V <- V + Ci * labs_i_sigma_sq
+        }
+    }
+    
+    # multiply by overall scale
     V <- V * sigma_sq
+    
     # return
     return( V )
 }