+++ R/sim_trait_mvn.R
@@ -4,13 +4,10 @@
 #' Although these traits have the covariance structure of genetic traits, and have heritabilities that can be estimated, they do not have causal loci (an association test against any locus should fail).
 #' Below `n` is the number of individuals.
 #'
+#' @inheritParams sim_trait
 #' @param rep The number of replicate traits to simulate.
 #' Simulating all you need at once is more efficient than simulating each separately (the kinship matrix is eigendecomposed once per run, shared across replicates).
 #' @param kinship The `n`-by-`n` kinship matrix of the individuals to simulate from.
-#' @param herit The desired heritability (proportion of trait variance due to genetics).
-#' @param mu The desired parametric mean value of the trait (default zero).
-#' @param sigma_sq The desired parametric variance factor of the trait (default 1).
-#' This factor corresponds to the variance of an outbred individual (see [cov_trait()]).
 #' @param tol Tolerance factor for an internal test of positive semi-definiteness of the trait covariance matrix.
 #' Procedure fails if any eigenvalues are smaller than `-tol` times the absolute value of the largest eigenvalue.
 #' Increase this value only if you are getting errors but you're sure your covariance matrix (the output of [cov_trait()]) is positive semi-definite.
@@ -32,6 +29,9 @@
 #' traits <- sim_trait_mvn( rep = 10, kinship = kinship, herit = 0.8 )
 #' traits
 #' 
+#' @seealso
+#' [cov_trait()], [sim_trait()]
+#' 
 #' @export
 sim_trait_mvn <- function(
                           rep,
@@ -39,6 +39,8 @@ sim_trait_mvn <- function(
                           herit,
                           mu = 0,
                           sigma_sq = 1,
+                          labs = NULL,
+                          labs_sigma_sq = NULL,
                           tol = 1e-06
                           ) {
     # check for missing parameters
@@ -61,11 +63,17 @@ sim_trait_mvn <- function(
     if (sigma_sq <= 0)
         stop('`sigma_sq` must be positive!')
 
+    # check labs with this shared function
+    n_ind <- nrow( kinship )
+    labs <- check_labs( labs, labs_sigma_sq, n_ind, herit )
+
     # construct the total covariance matrix of the trait
     V <- cov_trait(
         kinship = kinship,
         herit = herit,
-        sigma_sq = sigma_sq
+        sigma_sq = sigma_sq,
+        labs = labs,
+        labs_sigma_sq = labs_sigma_sq
     )
 
     # calculate matrix square root, such that
@@ -73,7 +81,6 @@ sim_trait_mvn <- function(
     V_sqrt <- sqrt_matrix( V, tol = tol )
 
     # simulate large matrix with IID standard normal values
-    n_ind <- nrow( kinship )
     traits <- matrix(
         stats::rnorm( rep * n_ind ),
         nrow = rep,